+++ cmd/mail.go
@@ -5,7 +5,7 @@ import "github.com/spf13/cobra"
 func newMailCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "mail",
-		Short: "Manage outbound mail (SES identities, inboxes, send)",
+		Short: "Email sending and inboxes on your domains",
 	}
 	cmd.AddCommand(newMailDomainsCmd(), newMailInboxesCmd(), newMailSendCmd(), newMailMessagesCmd())
 	return cmd